daemon: extract pam_auth module from config.rs
Partial close on #41. The full 6-way split suggested in the ticket
(schema/io/migration/secrets/auth/views) is mechanical but
cross-references between sections make a single-PR full split risky.
Started with the most self-contained slice — the inline
`#[cfg(target_os = "linux")] mod pam_auth { ... }` block at the bottom
of config.rs is already its own module and lifts cleanly into a file.

Effect: config.rs goes from 4373 → 4144 lines; pam_auth lands in its
own 241-line file. The PAM FFI is also the most security-sensitive
code in the daemon, so isolating it has independent value beyond
"just split the file".

`config.rs` is now `config/mod.rs`; the relative `include_str!` paths
in the migration tests gain an extra `../` for the new depth.

Followups for the rest of the split (most likely candidates):
- `config/secrets.rs` for the secret backends (KWallet, keyring,
  systemd-creds, file). These functions tightly cross-reference each
  other but are otherwise self-contained.
- `config/jwt.rs` for `generate_ws_jwt` / `validate_ws_jwt` /
  `OidcValidator`. JWT is shallow but couples with `read_common_file_secret`
  which lives in the (still-unsplit) secrets section.
